feat: Transient Data Dissemination
Implemented a dissemination plan for transient
data that is based on a deterministic algorithm.
The hash of the key is used first to determine the
org to which to disseminate and then the same hash
is used to select the peer within the org. The number
of peers to which to disseminate is specific in the
collection policy by 'maxPeers'.

closes #54

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>