feat: Extract transient data R/W sets and persist
Add a transient data store that extracts transient data write-sets
created during endorsement and stores the data to the transient data
cache. The standard private data read/writes are stored to the standard
transient store, so that they be available at commit time.

Also, added directories mod/peer/endorser and mod/peer/statedb with
default implementations in order for the build to succeed.

closes #53

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>