feat: Retrieve Transient Data
Using a deterministic algorithm, retrieve transient data
from the local peer or, using Gossip, from other peers that
are supposed to contain the data.

The collection policy is checked before returning the data,
i.e. if the requesting peer is not authorized to 'see' the
data then nil is returned for the data request.

closes #55

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>