feat: Retrieve Off-Ledger data
Retrieve Off-Ledger data from the current peer. If the data does not exist
then a Gossip message is sent to other peers requesting the data. Once the
data is successfully retrieved, it is stored on the current peer.

Note that before retrieving the data, a check is made to see if the current
peer has access to the data (according to collection policy). If the peer
does not have access, then nil is returned.

closes #63

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>