fix: Fix for TxSimulator validation error on DCAS keys
A DCAS, Base64-encoded key will sometimes have an underscore character causing PutPrivateData to fail from within chaincode. The TxSimulator validation logic rejects keys with leading underscores (in the case where the state DB is CouchDB) before the DCAS decorator can encode it with Base58 encoding.
This pull request modifies the existing logic where keys are automatically encoded in Base58 by the decorator, leaving it up to the caller to perform the Base58 encoding. The keys are still validated in order to ensure that the key is the hash of the value.

closes #185

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>