fix: Don't store transient data on endorser if not assigned
This patch adds a check during persistence of transient data to make sure that the key is assigned to the peer. If not then it is not stored locally (but still disseminated to other peers).

Also, changed the name of the loggers for all sub-packages of transientdata to be consistent in order to help in debugging.

closes #211

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>