fix: No match for query if criteria is 'unique'
When a search is performed using criteria that is deemed to be 'complete' then the retrieval is performed by key as opposed to by indexed query.
In the case where applications have components, and an AppName and AppVersion were provided, then only a key-based retrieval was performed. This resulted in the key not being found.
This patch modifies the behavior by also performing an index-based query if the key retrieval resulted in nil.

closes #251

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>