fix: Off-ledger collection data should not be pushed to all committers
This patch ensures that all off-ledger (and DCAS) data is not automatically pushed (disseminated) to all committers but is pushed to number of committers specified by MaximumPeerCount in the collection policy. If MaximumPeerCount is greater than the number of committers then additional endorsers are selected.
Also, RequiredPeerCount/MaximumPeerCount are not required to be greater than 0. In some situations you just want to store the data to the local peer. Although, if MaximumPeerCount is set to 0 but the local peer is not a committer then a random committer is chosen to dissminate to.

close #266

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>