feat: Implemented in-process user chaincode
Added a new type of chaincode that is similar to user chaincode but runs in the peer's process. In this case the chaincode does not need to be installed (since it is built with the peer) but needs to be instantiated with an endorsement and (optional) collection policy.

In order to demonstrate in-process user chaincode, all test chaincodes for the off-ledger and transient-data BDD tests were converted to in-process chaincodes.

closes #258

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>