fix: Register legacy and in-process changecodes on startup
After a restart, legacy chaincodes were not being registered since the chaincode directory path was not set. Also, in-process chaincodes were not being registered.
This patch sets the legacy chaincode path and adds all in-process chaincodes to the CC meta data.
The integration tests have been modified to perform a restart of all peers in order to demonstrate that user chaincodes and in-process user chaincodes are re-registered on peer startup.

Also, point to a new version of the SDK that has a couple of bug fixes that will make the tests more stable.

closes #328

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>