fix: Committer should not reject off-ledger data requests
When a local peer does not have the off-ledger data stored then it asks for the data from other peers using Gossip. Non-endorsing committers are also being asked for the data but the non-endorser was rejecting the request. This patch changes this logic so that a committing peer will also respond with the data since all endorsers may be unavailable.

closes #327

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>