feat: Added versioning to in-process user chaincode
In-process user chaincode may now specify a version such that multiple versions of the same chaincode may coexist.

A chaincode is registered with a version that consists of a major version and optional minor version, e.g. v1, v1.0. A chaincode may be instantiated and/or upgraded with an optional revision, e.g. v1.0.1, but the revision is not used when resolving the chaincode implementation but only as a means by which to update a chaincode's policies.

close #349

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>