feat: Enable orgs to share Couch DB instance
Added a PartitionType peer config that allows databases to be partioned by peer ID or MSP, allowing multiple peers to share the same CouchDB instance/cluster.

In non-clustered mode (i.e. each peer has a dedicated database), PartitionType is set to PEER which results in the peer ID being prefixed to the name of the database. In clustered mode (i.e. all peers in the same org share a database) the PartitionType is set to MSP which causes the database name to be prefixed with the MSP ID.

closes #348

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>