feat: Get DB artifacts for specified collections
The signature of function, GetDBArtifacts, was changed such that the names of the collections for the chaincode are passed in as an argument. The in-process user chaincode decides which indexes should be returned for each collection. This allows names of the collections and the number of collections to be dynamic.

closes #379

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>