fix: Enable cache for DCAS collections
Off-ledger collections cannot reliably cache data since updates to a key may not be propagated to all peers and a subsequent Get may retrieve stale data from the cache. Although, DCAS collections can reliably cache data, since an update to a DCAS value will always result in the same value. Deletes are still a problem since a DCAS value may still be cached on one or more peers. Support for delete of a DCAS key should be deprecated.

This cmmit (by default) enables the cache for DCAS collection types and disables the cache for off-ledger collection types. (Note that the cache may still be disabled for DCAS collection types by setting the off-ledger cache size to 0.)

This commit also updates to the latest fabric-sdk-go.

closes #382

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>