feat: Added option to always persist transient data
Previously transient data was only stored in memory unless the maximum cache size had been exceeded, at which time the oldest entries were offloaded to a database. This commit adds a configuration parameter to allow transient data to always be persisted (as well as cached). This way, if a peer goes down temporarily then the transient data will still be available when it comes back up.

closes #520

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>