feat: Use IPFS CIDs in DCAS
The DCAS is now using IPFS libraries to properly encode the content and generate standard CIDs. DCAS now supports storing objects and files.

Objects may be stored with the following encodings:

- Raw - The data is stored exactly as input by the client without trying to interpret the data. (This is the default encoding.)
- CBOR - Concise Binary Object Representation (or dag-cbor)
- Protobuf - The protobuf (or dag-pb) encoding contains data and links to other data

File data is either stored as a raw node or, if it's a large file then it is broken up into chunks and stored as a Merkle DAG.

closes #526

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>