fix: Non-committing peer returns error when handling private data state request
When a non-committing peer receives a Gossip request for private data for a specific block it would return an error that the last committed block was not at the desired height. This was because the 'lastCommittedBlockNum' is cached and only updated in the Commit function. Since Commit is never called on a non-committing peer, the value for 'lastCommittedBlockNum' was stuck at the value it was set to when the peer was started.

This commit registers a block handler with the Block Publisher in the CouchDB private data store to update 'lastCommittedBlockNum' when it receives a published block.

closes #547

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>